Deduplicate Video role definitions across Event, World, and migration.
Share default_roles/default_grants from core.permissions, derive SYSTEM_ROLES
and legacy JWT aliases from VIDEO_ROLE_PERMISSIONS maps, and reuse those
maps in migration 0050.
