+++ app/eventyay/core/permissions.py
@@ -93,9 +93,50 @@ VIDEO_CONFIG_MANAGER_PERMISSIONS = [
     Permission.EVENT_UPDATE.value,
 ]
 
+VIDEO_ROLE_PERMISSIONS: dict[str, list[str]] = {
+    'video_content_manager': VIDEO_CONTENT_MANAGER_PERMISSIONS,
+    'video_moderator': VIDEO_MODERATOR_PERMISSIONS,
+    'video_kiosk_manager': VIDEO_KIOSK_MANAGER_PERMISSIONS,
+    'video_analyst': VIDEO_ANALYST_PERMISSIONS,
+    'video_config_manager': VIDEO_CONFIG_MANAGER_PERMISSIONS,
+}
+
+# Pre-consolidation roles kept for in-flight JWTs / trait grants.
+LEGACY_VIDEO_ROLE_PERMISSIONS: dict[str, list[str]] = {
+    'video_stage_manager': [
+        Permission.EVENT_ROOMS_CREATE_STAGE.value,
+    ],
+    'video_channel_manager': [
+        Permission.EVENT_ROOMS_CREATE_CHAT.value,
+        Permission.EVENT_ROOMS_CREATE_BBB.value,
+    ],
+    'video_announcement_manager': [
+        Permission.EVENT_ANNOUNCE.value,
+    ],
+    'video_user_viewer': [
+        Permission.EVENT_USERS_LIST.value,
+    ],
+    'video_user_moderator': [
+        Permission.EVENT_USERS_MANAGE.value,
+        Permission.ROOM_CHAT_MODERATE.value,
+    ],
+    'video_room_manager': [
+        Permission.ROOM_UPDATE.value,
+        Permission.ROOM_DELETE.value,
+    ],
+    'video_poll_question_manager': [
+        Permission.ROOM_QUESTION_READ.value,
+        Permission.ROOM_QUESTION_MODERATE.value,
+        Permission.ROOM_POLL_READ.value,
+        Permission.ROOM_POLL_MANAGE.value,
+        Permission.ROOM_POLL_EARLY_RESULTS.value,
+    ],
+}
+
+LEGACY_VIDEO_ROLE_NAMES: tuple[str, ...] = tuple(LEGACY_VIDEO_ROLE_PERMISSIONS)
 
 SYSTEM_ROLES = {
-    "__kiosk": [
+    '__kiosk': [
         Permission.EVENT_VIEW.value,
         Permission.ROOM_VIEW.value,
         Permission.ROOM_CHAT_READ.value,
@@ -105,10 +146,10 @@ SYSTEM_ROLES = {
         Permission.ROOM_VIEWERS.value,
         Permission.ROOM_INVITE_ANONYMOUS.value,
     ],
-    "__anonymous_event": [
+    '__anonymous_event': [
         Permission.EVENT_VIEW.value,
     ],
-    "__anonymous_room": [
+    '__anonymous_room': [
         Permission.ROOM_QUESTION_READ.value,
         Permission.ROOM_QUESTION_ASK.value,
         Permission.ROOM_QUESTION_VOTE.value,
@@ -116,72 +157,117 @@ SYSTEM_ROLES = {
         Permission.ROOM_POLL_VOTE.value,
         Permission.ROOM_VIEW.value,
     ],
-    "video_content_manager": VIDEO_CONTENT_MANAGER_PERMISSIONS,
-    "video_moderator": VIDEO_MODERATOR_PERMISSIONS,
-    "video_kiosk_manager": VIDEO_KIOSK_MANAGER_PERMISSIONS,
-    "video_analyst": VIDEO_ANALYST_PERMISSIONS,
-    "video_config_manager": VIDEO_CONFIG_MANAGER_PERMISSIONS,
-    # Legacy role names kept so in-flight JWTs minted before consolidation still work.
-    "video_stage_manager": [
-        Permission.EVENT_ROOMS_CREATE_STAGE.value,
-    ],
-    "video_channel_manager": [
-        Permission.EVENT_ROOMS_CREATE_CHAT.value,
-        Permission.EVENT_ROOMS_CREATE_BBB.value,
-    ],
-    "video_announcement_manager": [
-        Permission.EVENT_ANNOUNCE.value,
-    ],
-    "video_user_viewer": [
-        Permission.EVENT_USERS_LIST.value,
-    ],
-    "video_user_moderator": [
-        Permission.EVENT_USERS_MANAGE.value,
-        Permission.ROOM_CHAT_MODERATE.value,
-    ],
-    "video_room_manager": [
-        Permission.ROOM_UPDATE.value,
-        Permission.ROOM_DELETE.value,
-    ],
-    "video_poll_question_manager": [
-        Permission.ROOM_QUESTION_READ.value,
-        Permission.ROOM_QUESTION_MODERATE.value,
-        Permission.ROOM_POLL_READ.value,
-        Permission.ROOM_POLL_MANAGE.value,
-        Permission.ROOM_POLL_EARLY_RESULTS.value,
-    ],
+    **{name: list(perms) for name, perms in VIDEO_ROLE_PERMISSIONS.items()},
+    **{name: list(perms) for name, perms in LEGACY_VIDEO_ROLE_PERMISSIONS.items()},
 }
 
 # Roles that are considered organizer/admin roles for permission management
-ORGANIZER_ROLES = frozenset({
-    'admin',
-    'apiuser',
-    'scheduleuser',
-    'video_content_manager',
-    'video_moderator',
-    'video_kiosk_manager',
-    'video_analyst',
-    'video_config_manager',
-    # Legacy aliases
-    'video_stage_manager',
-    'video_channel_manager',
-    'video_announcement_manager',
-    'video_user_viewer',
-    'video_user_moderator',
-    'video_room_manager',
-    'video_poll_question_manager',
-})
+ORGANIZER_ROLES = frozenset(
+    {
+        'admin',
+        'apiuser',
+        'scheduleuser',
+        *VIDEO_ROLE_PERMISSIONS,
+        *LEGACY_VIDEO_ROLE_PERMISSIONS,
+    }
+)
+
+
+def default_roles():
+    """Shared Event/World role → permission map (single source of truth)."""
+    attendee = [
+        Permission.EVENT_VIEW,
+        Permission.EVENT_EXHIBITION_CONTACT,
+        Permission.EVENT_CHAT_DIRECT,
+    ]
+    viewer = attendee + [Permission.ROOM_VIEW, Permission.ROOM_CHAT_READ]
+    participant = viewer + [
+        Permission.ROOM_CHAT_JOIN,
+        Permission.ROOM_CHAT_SEND,
+        Permission.ROOM_QUESTION_READ,
+        Permission.ROOM_QUESTION_ASK,
+        Permission.ROOM_QUESTION_VOTE,
+        Permission.ROOM_POLL_READ,
+        Permission.ROOM_POLL_VOTE,
+        Permission.ROOM_ROULETTE_JOIN,
+        Permission.ROOM_BBB_JOIN,
+        Permission.ROOM_JANUSCALL_JOIN,
+        Permission.ROOM_ZOOM_JOIN,
+    ]
+    room_creator = [Permission.EVENT_ROOMS_CREATE_CHAT]
+    room_owner = participant + [
+        Permission.ROOM_INVITE,
+        Permission.ROOM_DELETE,
+    ]
+    speaker = participant + [
+        Permission.ROOM_BBB_MODERATE,
+        Permission.ROOM_JANUSCALL_MODERATE,
+        Permission.ROOM_POLL_EARLY_RESULTS,
+    ]
+    moderator = speaker + [
+        Permission.ROOM_VIEWERS,
+        Permission.ROOM_CHAT_MODERATE,
+        Permission.ROOM_ANNOUNCE,
+        Permission.ROOM_BBB_RECORDINGS,
+        Permission.ROOM_QUESTION_MODERATE,
+        Permission.ROOM_POLL_EARLY_RESULTS,
+        Permission.ROOM_POLL_MANAGE,
+        Permission.EVENT_ANNOUNCE,
+    ]
+    admin = (
+        moderator
+        + room_creator
+        + [
+            Permission.EVENT_UPDATE,
+            Permission.ROOM_DELETE,
+            Permission.ROOM_UPDATE,
+            Permission.ROOM_INVITE,
+            Permission.EVENT_ROOMS_CREATE_BBB,
+            Permission.EVENT_ROOMS_CREATE_STAGE,
+            Permission.EVENT_ROOMS_CREATE_EXHIBITION,
+            Permission.EVENT_ROOMS_CREATE_POSTER,
+            Permission.EVENT_USERS_LIST,
+            Permission.EVENT_USERS_MANAGE,
+            Permission.EVENT_KIOSKS_MANAGE,
+            Permission.EVENT_GRAPHS,
+            Permission.EVENT_CONNECTIONS_UNLIMITED,
+        ]
+    )
+    apiuser = admin + [Permission.EVENT_API, Permission.EVENT_SECRETS]
+    scheduleuser = [Permission.EVENT_API]
+    roles = {
+        'attendee': attendee,
+        'viewer': viewer,
+        'participant': participant,
+        'room_creator': room_creator,
+        'room_owner': room_owner,
+        'speaker': speaker,
+        'moderator': moderator,
+        'admin': admin,
+        'apiuser': apiuser,
+        'scheduleuser': scheduleuser,
+    }
+    roles.update({name: list(perms) for name, perms in VIDEO_ROLE_PERMISSIONS.items()})
+    return roles
+
+
+def default_grants():
+    return {
+        'attendee': ['attendee'],
+        'admin': ['admin'],
+        'scheduleuser': [],
+    }
 
 
 def normalize_permission_value(permission):
     """Normalize permission to string value for comparison.
-    
+
     Args:
         permission: Permission enum or string value
-        
+
     Returns:
         str: Permission value as string
-        
+
     Raises:
         TypeError: If permission is not a string or Permission enum
     """
@@ -196,17 +282,17 @@ def normalize_permission_value(permission):
 
 def traits_match_required(traits: list[str], required_traits: list) -> bool:
     """Check if user traits match required traits for a role.
-    
+
     Args:
         traits: List of user traits
         required_traits: List of required traits (can contain nested lists for OR logic)
-        
+
     Returns:
         bool: True if all required traits are matched
     """
     if not isinstance(required_traits, list):
         return False
-    
+
     return all(
         any(x in traits for x in (r if isinstance(r, list) else [r]))
         for r in required_traits
